Update script.js
Here, we're creating two Tone.js Player objects, one for each audio file. We're using the toDestination() method to route the audio to the default output of the Web Audio API, and then using the Tone.Panner object to pan each audio file to the left or right channel.
In this example, we're panning "SongA.mp3" to the right channel (new Tone.Panner(1)) and "SongB.mp3" to the left channel (new Tone.Panner(-1)).
Then, we're updating the event listeners to start and stop the Tone.js Player objects instead of directly playing and pausing the audio elements.